requirements.txt
django-crispy-forms
django-debug-toolbar
django-dotenv
crispy-bootstrap4
sorl-thumbnail
django-cors-headers
django-filter
django-allauth
django_redis
Django==4.2
django_filter==23.1
django_hitcount==1.3.5
django_money==3.1.0
django_phonenumber_field==7.1.0
django-phonenumbers
djangorestframework==3.14.0
djangorestframework_simplejwt==5.2.2
Pillow==9.5.0
Pillow==9.5.0
selenium==4.9.1
